requirements.txt
yt_dlp
moviepy
ffmpeg
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech
spleeter